feat: 충전 한도 타입 수정 - 이후 확장성을 위해 int에서 long으로 변경

+++ src/main/java/org/c4marathon/assignment/entity/Account.java
@@ -57,7 +57,7 @@ public class Account extends BaseEntity {
         this.dailyChargeAmountUpdatedDate = Instant.now();
     }
 
-    public boolean isDailyLimitExceeded(int amount) {
+    public boolean isDailyLimitExceeded(long amount) {
         LocalDate updatedDate = this.dailyChargeAmountUpdatedDate.atZone(ZoneId.systemDefault()).toLocalDate();
         LocalDate now = Instant.now().atZone(ZoneId.systemDefault()).toLocalDate();
